+++ package.json
@@ -2,6 +2,8 @@
   "name": "kekstagram",
   "version": "26.0.0",
   "private": true,
+  "main": "main.js",
+  "type": "module",
   "description": "Личный проект «Кекстаграм» от HTML Academy",
   "repository": {
     "type": "git",
